New utility union_merge_observations
New data manipulation to support a flagged revisions experiment.  This
will merge two sets of observations so that any rows that match a row
in another file by ID are collapsed into a single row in the output.
We assume that IDs never repeat in a single file (if they do, we drop
all but the last occurrence).  The rows are collapsed by doing a set
union of the properties, with matching properties from later input
files taking precedent.

TODO:
* Not optimized, demands all the memory.

Bug: https://phabricator.wikimedia.org/T166235