Improve error message when decoding a non-item revision from Wikidata
Revscoring returns a very verbose and cryptic JSONDecodeError stacktrace
when it tries to score a non-item Wikidata revision with the item
quality model (for example, a User/Talk page change, etc..).
This change adds a new exception meant to capture this kind of error,
to wrap it in a clearer error message.

Co-authored-by: halfak <aaron.halfaker@gmail.com>,
Aiko Chou <achou@wikimedia.org>
Bug: T302851