Release revscoring 2.11.3
The release 2.11.2 should be the right one, but I realized right now
that I published the artifact from my local repo that may have been
messed up (limited to the new code that Aiko wrote).

To avoid any issue while deploying the new code, I propose to bump
the version of revscoring again, to re-publish a new artifact to Pypi.

Bug: T302851