extractors: Add a simple MWAPI HTTP cache to the api extractor
The Wikimedia ML team is porting revscoring to the Kserve
architecture, that relies heavily on Tornado and async
processing. The API extractor uses the mwapi package behind
the scenes, that is in turn using blocking HTTP clients
to fetch data from the MediaWiki API. We are adding
support for async to mwapi in T309623, but due to
the nature of asyncio and Python co-routines, it would
require an invasive change to revscoring's api extractor
to support both blocking and non-blocking HTTP calls.

In this change the proposal it to simply add a proxy
object called MWAPICache, that can be passed as
argument to the extract() function and to the Extractor's init.
The object can be populated asyncronously by
the revscoring user via whatever async/non-blocking library
is preferred, to untagle the dependency to the (blocking)
mwapi Session.

Bug: T309623