ci: use bullseye image in CI (#531)
Using bullseye image for CI. In this PR we install the additional dependencies required to run tests etc using Debian Bullseye which is used by inference-services.
Changes:

CI commands exist in Makefile so that we can easily transfer them to any platform
updated dictionaries that have a hunspell package in debian. hunspell is a "superset" of myspell. Will also change these in inference-services image.
fixed failing tests due to changes in language dictionaries
added np.random.seed for test that failed randomly due to the values hardcoded in the test
skipped test that is failing in CI for serbian language - to revisit with the python upgrade